LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/require-corp-sw-from-none.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt?passthrough because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt?passthrough due to access control checks.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html.

PASS setting up
PASS making a same-origin request for CORP: same-origin
PASS making a same-origin request for no CORP
PASS making a same-origin request for CORP: cross-origin
PASS making a cross-origin request for CORP: same-origin
FAIL making a cross-origin request for no CORP assert_unreached: Should have rejected: undefined Reached unreachable code
PASS making a cross-origin request for CORP: cross-origin
PASS making a cross-origin request for CORP: same-origin [PASS THROUGH]
PASS making a cross-origin request for no CORP [PASS THROUGH]
PASS making a cross-origin request for CORP: cross-origin [PASS THROUGH]
PASS making a cross-origin request with CORS without ACAO
PASS making a cross-origin request with CORS
